move to functions-array
