+++ datafusion/sql/src/expr/mod.rs
@@ -18,9 +18,9 @@
 use arrow_schema::DataType;
 use arrow_schema::TimeUnit;
 use datafusion_common::exec_err;
-use datafusion_common::utils::list_ndims;
+use datafusion_expr::planner::FieldAccessExpr;
+use datafusion_expr::planner::PlannerSimplifyResult;
 use sqlparser::ast::{CastKind, Expr as SQLExpr, Subscript, TrimWhereField, Value};
-use std::sync::Arc;
 
 use datafusion_common::{
     internal_datafusion_err, internal_err, not_impl_err, plan_err, DFSchema, Result,
@@ -29,12 +29,11 @@ use datafusion_common::{
 use datafusion_expr::expr::InList;
 use datafusion_expr::expr::ScalarFunction;
 use datafusion_expr::{
-    lit, AggregateFunction, Between, BinaryExpr, Cast, Expr, ExprSchemable,
-    GetFieldAccess, Like, Literal, Operator, ScalarUDF, TryCast,
+    lit, Between, BinaryExpr, Cast, Expr, ExprSchemable, GetFieldAccess, Like, Literal,
+    Operator, TryCast,
 };
 
-use crate::planner::PlannerSimplifyResult;
-use crate::planner::{ContextProvider, PlannerContext, SqlToRel, UserDefinedPlanner};
+use crate::planner::{ContextProvider, PlannerContext, SqlToRel};
 
 mod binary_op;
 mod function;
@@ -108,22 +107,11 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
         schema: &DFSchema,
     ) -> Result<Expr> {
         // try extension planers
-        let mut binary_expr = crate::planner::BinaryExpr { op, left, right };
-        let num_planners = self.planners.len();
-        for (i, planner) in self.planners.iter().enumerate() {
+        let mut binary_expr = datafusion_expr::planner::BinaryExpr { op, left, right };
+        for planner in self.planners.iter() {
             match planner.plan_binary_op(binary_expr, schema)? {
                 PlannerSimplifyResult::Simplified(expr) => {
                     return Ok(expr);
-                }
-                PlannerSimplifyResult::OriginalBinaryExpr(expr)
-                    if i + 1 == num_planners =>
-                {
-                    let crate::planner::BinaryExpr { op, left, right } = expr;
-                    return Ok(Expr::BinaryExpr(BinaryExpr::new(
-                        Box::new(left),
-                        self.parse_sql_binary_op(op)?,
-                        Box::new(right),
-                    )));
                 }
                 PlannerSimplifyResult::OriginalBinaryExpr(expr) => {
                     binary_expr = expr;
@@ -134,7 +122,12 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
             }
         }
 
-        internal_err!("Unexpect to reach here")
+        let datafusion_expr::planner::BinaryExpr { op, left, right } = binary_expr;
+        Ok(Expr::BinaryExpr(BinaryExpr::new(
+            Box::new(left),
+            self.parse_sql_binary_op(op)?,
+            Box::new(right),
+        )))
     }
 
     /// Generate a relational expression from a SQL expression
@@ -289,16 +282,11 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
                     }
                 };
 
-                let mut field_access_expr =
-                    crate::planner::FieldAccessExpr { expr, field_access };
-                let num_planners = self.planners.len();
-                for (i, planner) in self.planners.iter().enumerate() {
+                let mut field_access_expr = FieldAccessExpr { expr, field_access };
+                for planner in self.planners.iter() {
                     match planner.plan_field_access(field_access_expr, schema)? {
                         PlannerSimplifyResult::Simplified(expr) => {
                             return Ok(expr)
-                        }
-                        PlannerSimplifyResult::OriginalFieldAccessExpr(_) if i + 1 == num_planners => {
-                            return internal_err!("Expected a simplified result, but none was found")
                         }
                         PlannerSimplifyResult::OriginalFieldAccessExpr(expr) => {
                             field_access_expr = expr;
@@ -309,7 +297,7 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
                     }
                 }
 
-                internal_err!("Unexpect to reach here")
+                internal_err!("Expected a simplified result, but none was found")
             }
 
             SQLExpr::CompoundIdentifier(ids) => {
@@ -929,210 +917,6 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
         let args = vec![fullstr, substr];
         Ok(Expr::ScalarFunction(ScalarFunction::new_udf(fun, args)))
     }
-}
-
-pub struct ArrayFunctionPlanner {
-    array_concat: Arc<ScalarUDF>,
-    array_append: Arc<ScalarUDF>,
-    array_prepend: Arc<ScalarUDF>,
-    array_has_all: Arc<ScalarUDF>,
-    make_array: Arc<ScalarUDF>,
-}
-
-impl ArrayFunctionPlanner {
-    pub fn try_new(context_provider: &dyn ContextProvider) -> Result<Self> {
-        let Some(array_concat) = context_provider.get_function_meta("array_concat")
-        else {
-            return internal_err!("array_concat not found");
-        };
-        let Some(array_append) = context_provider.get_function_meta("array_append")
-        else {
-            return internal_err!("array_append not found");
-        };
-        let Some(array_prepend) = context_provider.get_function_meta("array_prepend")
-        else {
-            return internal_err!("array_prepend not found");
-        };
-        let Some(array_has_all) = context_provider.get_function_meta("array_has_all")
-        else {
-            return internal_err!("array_has_all not found");
-        };
-        let Some(make_array) = context_provider.get_function_meta("make_array") else {
-            return internal_err!("make_array not found");
-        };
-
-        Ok(Self {
-            array_concat,
-            array_append,
-            array_prepend,
-            array_has_all,
-            make_array,
-        })
-    }
-}
-impl UserDefinedPlanner for ArrayFunctionPlanner {
-    fn plan_binary_op(
-        &self,
-        expr: crate::planner::BinaryExpr,
-        schema: &DFSchema,
-    ) -> Result<PlannerSimplifyResult> {
-        let crate::planner::BinaryExpr { op, left, right } = expr;
-
-        if op == sqlparser::ast::BinaryOperator::StringConcat {
-            let left_type = left.get_type(schema)?;
-            let right_type = right.get_type(schema)?;
-            let left_list_ndims = list_ndims(&left_type);
-            let right_list_ndims = list_ndims(&right_type);
-
-            // Rewrite string concat operator to function based on types
-            // if we get list || list then we rewrite it to array_concat()
-            // if we get list || non-list then we rewrite it to array_append()
-            // if we get non-list || list then we rewrite it to array_prepend()
-            // if we get string || string then we rewrite it to concat()
-
-            // We determine the target function to rewrite based on the list n-dimension, the check is not exact but sufficient.
-            // The exact validity check is handled in the actual function, so even if there is 3d list appended with 1d list, it is also fine to rewrite.
-            if left_list_ndims + right_list_ndims == 0 {
-                // TODO: concat function ignore null, but string concat takes null into consideration
-                // we can rewrite it to concat if we can configure the behaviour of concat function to the one like `string concat operator`
-            } else if left_list_ndims == right_list_ndims {
-                return Ok(PlannerSimplifyResult::Simplified(
-                    self.array_concat.call(vec![left, right]),
-                ));
-            } else if left_list_ndims > right_list_ndims {
-                return Ok(PlannerSimplifyResult::Simplified(
-                    self.array_append.call(vec![left, right]),
-                ));
-            } else if left_list_ndims < right_list_ndims {
-                return Ok(PlannerSimplifyResult::Simplified(
-                    self.array_prepend.call(vec![left, right]),
-                ));
-            }
-        } else if matches!(
-            op,
-            sqlparser::ast::BinaryOperator::AtArrow
-                | sqlparser::ast::BinaryOperator::ArrowAt
-        ) {
-            let left_type = left.get_type(schema)?;
-            let right_type = right.get_type(schema)?;
-            let left_list_ndims = list_ndims(&left_type);
-            let right_list_ndims = list_ndims(&right_type);
-            // if both are list
-            if left_list_ndims > 0 && right_list_ndims > 0 {
-                if op == sqlparser::ast::BinaryOperator::AtArrow {
-                    // array1 @> array2 -> array_has_all(array1, array2)
-                    return Ok(PlannerSimplifyResult::Simplified(
-                        self.array_has_all.call(vec![left, right]),
-                    ));
-                } else {
-                    // array1 <@ array2 -> array_has_all(array2, array1)
-                    return Ok(PlannerSimplifyResult::Simplified(
-                        self.array_has_all.call(vec![right, left]),
-                    ));
-                }
-            }
-        }
-
-        Ok(PlannerSimplifyResult::OriginalBinaryExpr(
-            crate::planner::BinaryExpr { op, left, right },
-        ))
-    }
-
-    fn plan_array_literal(
-        &self,
-        exprs: Vec<Expr>,
-        _schema: &DFSchema,
-    ) -> Result<PlannerSimplifyResult> {
-        Ok(PlannerSimplifyResult::Simplified(
-            self.make_array.call(exprs),
-        ))
-    }
-}
-
-pub struct FieldAccessPlanner {
-    get_field: Arc<ScalarUDF>,
-    array_element: Arc<ScalarUDF>,
-    array_slice: Arc<ScalarUDF>,
-}
-
-impl FieldAccessPlanner {
-    pub fn try_new(context_provider: &dyn ContextProvider) -> Result<Self> {
-        let Some(get_field) = context_provider.get_function_meta("get_field") else {
-            return internal_err!("get_feild not found");
-        };
-        let Some(array_element) = context_provider.get_function_meta("array_element")
-        else {
-            return internal_err!("array_element not found");
-        };
-        let Some(array_slice) = context_provider.get_function_meta("array_slice") else {
-            return internal_err!("array_slice not found");
-        };
-
-        Ok(Self {
-            get_field,
-            array_element,
-            array_slice,
-        })
-    }
-}
-
-impl UserDefinedPlanner for FieldAccessPlanner {
-    fn plan_field_access(
-        &self,
-        expr: crate::planner::FieldAccessExpr,
-        _schema: &DFSchema,
-    ) -> Result<PlannerSimplifyResult> {
-        let crate::planner::FieldAccessExpr { expr, field_access } = expr;
-
-        match field_access {
-            // expr["field"] => get_field(expr, "field")
-            GetFieldAccess::NamedStructField { name } => {
-                Ok(PlannerSimplifyResult::Simplified(
-                    self.get_field.call(vec![expr, lit(name)]),
-                ))
-            }
-            // expr[idx] ==> array_element(expr, idx)
-            GetFieldAccess::ListIndex { key: index } => {
-                match expr {
-                    // Special case for array_agg(expr)[index] to NTH_VALUE(expr, index)
-                    Expr::AggregateFunction(agg_func) if is_array_agg(&agg_func) => {
-                        Ok(PlannerSimplifyResult::Simplified(Expr::AggregateFunction(
-                            datafusion_expr::expr::AggregateFunction::new(
-                                AggregateFunction::NthValue,
-                                agg_func
-                                    .args
-                                    .into_iter()
-                                    .chain(std::iter::once(*index))
-                                    .collect(),
-                                agg_func.distinct,
-                                agg_func.filter,
-                                agg_func.order_by,
-                                agg_func.null_treatment,
-                            ),
-                        )))
-                    }
-                    _ => Ok(PlannerSimplifyResult::Simplified(
-                        self.array_element.call(vec![expr, *index]),
-                    )),
-                }
-            }
-            // expr[start, stop, stride] ==> array_slice(expr, start, stop, stride)
-            GetFieldAccess::ListRange {
-                start,
-                stop,
-                stride,
-            } => Ok(PlannerSimplifyResult::Simplified(
-                self.array_slice.call(vec![expr, *start, *stop, *stride]),
-            )),
-        }
-    }
-}
-
-fn is_array_agg(agg_func: &datafusion_expr::expr::AggregateFunction) -> bool {
-    agg_func.func_def
-        == datafusion_expr::expr::AggregateFunctionDefinition::BuiltIn(
-            AggregateFunction::ArrayAgg,
-        )
 }
 
 #[cfg(test)]
