+++ datafusion/sql/src/expr/value.rs
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::planner::{ContextProvider, PlannerContext, PlannerSimplifyResult, SqlToRel};
+use crate::planner::{ContextProvider, PlannerContext, SqlToRel};
 use arrow::compute::kernels::cast_utils::parse_interval_month_day_nano;
 use arrow::datatypes::DECIMAL128_MAX_PRECISION;
 use arrow_schema::DataType;
@@ -24,6 +24,7 @@ use datafusion_common::{
     ScalarValue,
 };
 use datafusion_expr::expr::{BinaryExpr, Placeholder};
+use datafusion_expr::planner::PlannerSimplifyResult;
 use datafusion_expr::{lit, Expr, Operator};
 use log::debug;
 use sqlparser::ast::{BinaryOperator, Expr as SQLExpr, Interval, Value};
@@ -144,16 +145,10 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
             .collect::<Result<Vec<_>>>()?;
 
         let mut exprs = values;
-        let num_planners = self.planners.len();
-        for (i, planner) in self.planners.iter().enumerate() {
+        for planner in self.planners.iter() {
             match planner.plan_array_literal(exprs, schema)? {
                 PlannerSimplifyResult::Simplified(expr) => {
                     return Ok(expr);
-                }
-                PlannerSimplifyResult::OriginalArray(_) if i + 1 == num_planners => {
-                    return internal_err!(
-                        "Expected a simplified result, but none was found"
-                    )
                 }
                 PlannerSimplifyResult::OriginalArray(values) => exprs = values,
                 _ => {
@@ -164,7 +159,7 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
             }
         }
 
-        internal_err!("Unexpect to reach here")
+        internal_err!("Expected a simplified result, but none was found")
     }
 
     /// Convert a SQL interval expression to a DataFusion logical plan
