+++ datafusion/sql/src/planner.rs
@@ -21,18 +21,15 @@ use std::sync::Arc;
 use std::vec;
 
 use arrow_schema::*;
-use datafusion_common::file_options::file_type::FileType;
 use datafusion_common::{
     field_not_found, internal_err, plan_datafusion_err, DFSchemaRef, SchemaError,
 };
-use datafusion_expr::{GetFieldAccess, WindowUDF};
+use datafusion_expr::planner::UserDefinedPlanner;
 use sqlparser::ast::TimezoneInfo;
 use sqlparser::ast::{ArrayElemTypeDef, ExactNumberInfo};
 use sqlparser::ast::{ColumnDef as SQLColumnDef, ColumnOption};
 use sqlparser::ast::{DataType as SQLDataType, Ident, ObjectName, TableAlias};
 
-use crate::expr::{ArrayFunctionPlanner, FieldAccessPlanner};
-use datafusion_common::config::ConfigOptions;
 use datafusion_common::TableReference;
 use datafusion_common::{
     not_impl_err, plan_err, unqualified_field_not_found, DFSchema, DataFusionError,
@@ -40,64 +37,11 @@ use datafusion_common::{
 };
 use datafusion_expr::logical_plan::{LogicalPlan, LogicalPlanBuilder};
 use datafusion_expr::utils::find_column_exprs;
-use datafusion_expr::TableSource;
-use datafusion_expr::{col, AggregateUDF, Expr, ScalarUDF};
+use datafusion_expr::{col, Expr};
 
 use crate::utils::make_decimal_type;
 
-/// The ContextProvider trait allows the query planner to obtain meta-data about tables and
-/// functions referenced in SQL statements
-pub trait ContextProvider {
-    /// Getter for a datasource
-    fn get_table_source(&self, name: TableReference) -> Result<Arc<dyn TableSource>>;
-
-    fn get_file_type(&self, _ext: &str) -> Result<Arc<dyn FileType>> {
-        not_impl_err!("Registered file types are not supported")
-    }
-
-    /// Getter for a table function
-    fn get_table_function_source(
-        &self,
-        _name: &str,
-        _args: Vec<Expr>,
-    ) -> Result<Arc<dyn TableSource>> {
-        not_impl_err!("Table Functions are not supported")
-    }
-
-    /// This provides a worktable (an intermediate table that is used to store the results of a CTE during execution)
-    /// We don't directly implement this in the logical plan's ['SqlToRel`]
-    /// because the sql code needs access to a table that contains execution-related types that can't be a direct dependency
-    /// of the sql crate (namely, the `CteWorktable`).
-    /// The [`ContextProvider`] provides a way to "hide" this dependency.
-    fn create_cte_work_table(
-        &self,
-        _name: &str,
-        _schema: SchemaRef,
-    ) -> Result<Arc<dyn TableSource>> {
-        not_impl_err!("Recursive CTE is not implemented")
-    }
-
-    /// Getter for a UDF description
-    fn get_function_meta(&self, name: &str) -> Option<Arc<ScalarUDF>>;
-    /// Getter for a UDAF description
-    fn get_aggregate_meta(&self, name: &str) -> Option<Arc<AggregateUDF>>;
-    /// Getter for a UDWF
-    fn get_window_meta(&self, name: &str) -> Option<Arc<WindowUDF>>;
-    /// Getter for system/user-defined variable type
-    fn get_variable_type(&self, variable_names: &[String]) -> Option<DataType>;
-
-    /// Get configuration options
-    fn options(&self) -> &ConfigOptions;
-
-    /// Get all user defined scalar function names
-    fn udf_names(&self) -> Vec<String>;
-
-    /// Get all user defined aggregate function names
-    fn udaf_names(&self) -> Vec<String>;
-
-    /// Get all user defined window function names
-    fn udwf_names(&self) -> Vec<String>;
-}
+pub use datafusion_expr::planner::ContextProvider;
 
 /// SQL parser options
 #[derive(Debug)]
@@ -237,63 +181,55 @@ impl PlannerContext {
     }
 }
 
-/// This trait allows users to customize the behavior of the SQL planner
-pub trait UserDefinedPlanner {
-    /// Plan the binary operation between two expressions, return None if not possible
-    fn plan_binary_op(
-        &self,
-        expr: BinaryExpr,
-        _schema: &DFSchema,
-    ) -> Result<PlannerSimplifyResult> {
-        Ok(PlannerSimplifyResult::OriginalBinaryExpr(expr))
-    }
+// /// This trait allows users to customize the behavior of the SQL planner
+// pub trait UserDefinedPlanner {
+//     /// Plan the binary operation between two expressions, return None if not possible
+//     fn plan_binary_op(
+//         &self,
+//         expr: BinaryExpr,
+//         _schema: &DFSchema,
+//     ) -> Result<PlannerSimplifyResult> {
+//         Ok(PlannerSimplifyResult::OriginalBinaryExpr(expr))
+//     }
 
-    /// Plan the field access expression, return None if not possible
-    fn plan_field_access(
-        &self,
-        expr: FieldAccessExpr,
-        _schema: &DFSchema,
-    ) -> Result<PlannerSimplifyResult> {
-        Ok(PlannerSimplifyResult::OriginalFieldAccessExpr(expr))
-    }
+//     /// Plan the field access expression, return None if not possible
+//     fn plan_field_access(
+//         &self,
+//         expr: FieldAccessExpr,
+//         _schema: &DFSchema,
+//     ) -> Result<PlannerSimplifyResult> {
+//         Ok(PlannerSimplifyResult::OriginalFieldAccessExpr(expr))
+//     }
 
-    fn plan_array_literal(
-        &self,
-        exprs: Vec<Expr>,
-        _schema: &DFSchema,
-    ) -> Result<PlannerSimplifyResult> {
-        Ok(PlannerSimplifyResult::OriginalArray(exprs))
-    }
-}
+//     fn plan_array_literal(
+//         &self,
+//         exprs: Vec<Expr>,
+//         _schema: &DFSchema,
+//     ) -> Result<PlannerSimplifyResult> {
+//         Ok(PlannerSimplifyResult::OriginalArray(exprs))
+//     }
+// }
 
-pub struct BinaryExpr {
-    pub op: sqlparser::ast::BinaryOperator,
-    pub left: Expr,
-    pub right: Expr,
-}
+// pub struct BinaryExpr {
+//     pub op: sqlparser::ast::BinaryOperator,
+//     pub left: Expr,
+//     pub right: Expr,
+// }
 
-pub struct FieldAccessExpr {
-    pub field_access: GetFieldAccess,
-    pub expr: Expr,
-}
+// pub struct FieldAccessExpr {
+//     pub field_access: GetFieldAccess,
+//     pub expr: Expr,
+// }
 
-pub enum PlannerSimplifyResult {
-    /// The function call was simplified to an entirely new Expr
-    Simplified(Expr),
-    /// the function call could not be simplified, and the arguments
-    /// are return unmodified.
-    OriginalBinaryExpr(BinaryExpr),
-    OriginalFieldAccessExpr(FieldAccessExpr),
-    OriginalArray(Vec<Expr>),
-}
-
-pub enum PlanSimplifyResult {
-    /// The function call was simplified to an entirely new Expr
-    Simplified(Expr),
-    /// the function call could not be simplified, and the arguments
-    /// are return unmodified.
-    Original(Vec<Expr>),
-}
+// pub enum PlannerSimplifyResult {
+//     /// The function call was simplified to an entirely new Expr
+//     Simplified(Expr),
+//     /// the function call could not be simplified, and the arguments
+//     /// are return unmodified.
+//     OriginalBinaryExpr(BinaryExpr),
+//     OriginalFieldAccessExpr(FieldAccessExpr),
+//     OriginalArray(Vec<Expr>),
+// }
 
 /// SQL query planner
 pub struct SqlToRel<'a, S: ContextProvider> {
@@ -322,10 +258,6 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
     /// Create a new query planner
     pub fn new_with_options(context_provider: &'a S, options: ParserOptions) -> Self {
         let normalize = options.enable_ident_normalization;
-        let array_planner =
-            Arc::new(ArrayFunctionPlanner::try_new(context_provider).unwrap()) as _;
-        let field_access_planner =
-            Arc::new(FieldAccessPlanner::try_new(context_provider).unwrap()) as _;
 
         SqlToRel {
             context_provider,
@@ -333,9 +265,6 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
             normalizer: IdentNormalizer::new(normalize),
             planners: vec![],
         }
-        // todo put this somewhere else
-        .with_user_defined_planner(array_planner)
-        .with_user_defined_planner(field_access_planner)
     }
 
     pub fn build_schema(&self, columns: Vec<SQLColumnDef>) -> Result<Schema> {
